clean up older uploadFile function, clean up action naming for assistant data operations
