+++ src/assistant/data/asstDataOperationsProvider.ts
@@ -7,10 +7,6 @@ import {
   HTTPHeaders,
 } from '../../pinecone-generated-ts-fetch/assistant_data';
 import {
-  MetricsApi,
-  Configuration as EvalConfiguration,
-} from '../../pinecone-generated-ts-fetch/assistant_evaluation';
-import {
   buildUserAgent,
   getFetch,
   queryParamsStringify,
@@ -24,7 +20,6 @@ export class AsstDataOperationsProvider {
   private readonly asstName: string;
   private asstHostUrl?: string;
   private asstDataOperations?: ManageAssistantsDataApi;
-  private metrics?: MetricsApi;
   private additionalHeaders?: HTTPHeaders;
 
   constructor(
@@ -64,32 +59,10 @@ export class AsstDataOperationsProvider {
     return this.asstHostUrl;
   }
 
-  provideMetrics() {
-    if (this.metrics) {
-      return this.metrics;
-    } else {
-      this.asstHostUrl = this.getAsstHostByRegion();
-      this.metrics = this.buildAssistEvalOperationsConfig();
-    }
-    return this.metrics;
-  }
-
-  getAsstHostByRegion() {
-    const { assistantRegion } = this.config;
-    let asstHostUrl = 'https://prod-1-data.ke.pinecone.io/assistant';
-
-    // if the region has been specified as 'eu' use that, otherwise default to 'us'
-    if (assistantRegion && assistantRegion.toLowerCase() === 'eu') {
-      asstHostUrl = 'https://prod-eu-data.ke.pinecone.io/assistant';
-    }
-    return asstHostUrl;
-  }
-
   buildAsstDataOperationsConfig() {
     const { apiKey } = this.config;
-    // Get the host URL from the singleton
     const hostUrl = this.asstHostUrl;
-    const headers = this.config.additionalHeaders || null;
+    const headers = this.additionalHeaders || null;
     const apiConfig: AssistantOperationsApiConfigurationParameters = {
       basePath: hostUrl,
       apiKey,
@@ -102,33 +75,7 @@ export class AsstDataOperationsProvider {
       fetchApi: getFetch(this.config),
       middleware,
     };
-    console.log('API CONFIG GETTING PASSED THROUGH TO MADA', apiConfig);
-    console.log(
-      'THIS IS THE FETCH API SPECIFICALLY FOR MADA',
-      apiConfig.fetchApi
-    );
 
     return new ManageAssistantsDataApi(new DataConfiguration(apiConfig));
-  }
-
-  buildAssistEvalOperationsConfig() {
-    const { apiKey } = this.config;
-
-    const hostUrl = this.getAsstHostByRegion();
-    const headers = this.config.additionalHeaders || null;
-    const apiConfig: AssistantOperationsApiConfigurationParameters = {
-      basePath: hostUrl,
-      apiKey,
-      queryParamsStringify,
-      headers: {
-        'User-Agent': buildUserAgent(this.config),
-        'X-Pinecone-Api-Version': X_PINECONE_API_VERSION,
-        ...headers,
-      },
-      fetchApi: getFetch(this.config),
-      middleware,
-    };
-
-    return new MetricsApi(new EvalConfiguration(apiConfig));
   }
 }
